fix: 일정 등록 기간이 지났을 때 타이머가 이상하게 표시되는 오류 해결
