+++ src/types/roomInfo.ts
@@ -1,6 +1,5 @@
 export interface TimerTypes {
   deadLine: string;
-  isTimeExpired: boolean;
 }
 
 export interface participant {
